moving block parsing logic to pxtlib
